# docs/openapi/openapi.yaml
...
openapi: 3.0.4
...
components:
  schemas:
...
    ConnectionType:
      enum:
      - HTTP
      - MQTT
      type: string
...
    CreateGpsSourceConfigDto:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/GpsSourceType"
        username:
          type: string
        password:
          type: string
        token:
          type: string
        deviceId:
          type: string
        userId:
          $ref: "#/components/schemas/UUID"
        connectionType:
          $ref: "#/components/schemas/ConnectionType"
        filterInaccurateData:
          type: boolean
        maxAllowedAccuracy:
          format: int32
          type: integer
        maxAllowedSpeed:
          format: int32
          type: integer
        enableDuplicateDetection:
          type: boolean
        duplicateDetectionThresholdMinutes:
          format: int32
          type: integer
...
    DawarichGeometry:
      type: object
      properties:
        type:
          type: string
        coordinates:
          type: array
          items:
            format: double
            type: number
        longitude:
          format: double
          type: number
        latitude:
          format: double
          type: number
    DawarichLocation:
      type: object
      properties:
        type:
          type: string
        geometry:
          $ref: "#/components/schemas/DawarichGeometry"
        properties:
          $ref: "#/components/schemas/DawarichProperties"
    DawarichPayload:
      type: object
      properties:
        locations:
          type: array
          items:
            $ref: "#/components/schemas/DawarichLocation"
    DawarichProperties:
      type: object
      properties:
        track_id:
          type: string
        course_accuracy:
          format: double
          type: number
        timestamp:
          $ref: "#/components/schemas/Instant"
        altitude:
          format: double
          type: number
        speed:
          format: double
          type: number
        horizontal_accuracy:
          format: double
          type: number
        vertical_accuracy:
          format: double
          type: number
        device_id:
          type: string
        speed_accuracy:
          format: double
          type: number
        course:
          format: double
          type: number
...
    GpsSourceType:
      enum:
      - OWNTRACKS
      - GPSLOGGER
      - OVERLAND
      - TRACCAR
      - GOOGLE_TIMELINE
      - GPX
      - DAWARICH
      - HOME_ASSISTANT
      - GEOJSON
      - CSV
      - COLOTA
      - MANUAL
      - MOBILE_APP
      type: string
...
    HomeAssistantBattery:
      type: object
      properties:
        level:
          format: int32
          type: integer
    HomeAssistantGpsData:
      type: object
      properties:
        device_id:
          type: string
        timestamp:
          $ref: "#/components/schemas/Instant"
        location:
          $ref: "#/components/schemas/HomeAssistantLocation"
        battery:
          $ref: "#/components/schemas/HomeAssistantBattery"
    HomeAssistantLocation:
      type: object
      properties:
        latitude:
          format: double
          type: number
        longitude:
          format: double
          type: number
        accuracy:
          format: double
          type: number
        altitude:
          format: double
          type: number
        speed:
          format: double
          type: number
...
    Instant:
      format: date-time
      type: string
      example: 2022-03-10T16:15:50Z
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...
    UpdateGpsSourceConfigDto:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        username:
          type: string
        password:
          type: string
        token:
          type: string
        deviceId:
          type: string
        userId:
          type: string
        connectionType:
          $ref: "#/components/schemas/ConnectionType"
        filterInaccurateData:
          type: boolean
        maxAllowedAccuracy:
          format: int32
          type: integer
        maxAllowedSpeed:
          format: int32
          type: integer
        enableDuplicateDetection:
          type: boolean
        duplicateDetectionThresholdMinutes:
          format: int32
          type: integer
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
paths:
...
  /api/dawarich/api/v1/points:
    post:
      summary: Ingest Dawarich points
      description: Receives Dawarich point payloads and stores them for the matching
        source token.
      operationId: DawarichResource_handleDawarichGet
      tags:
      - "Public: GPS Integrations: Dawarich"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DawarichPayload"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
  /api/dawarich/api/v1/stats:
    get:
      summary: Get Dawarich compatibility stats
      description: Returns a Dawarich-compatible stats response for authenticated
        compatibility clients.
      operationId: DawarichResource_handleDawarichStats
      tags:
      - "Public: GPS Integrations: Dawarich"
      parameters:
      - name: api_key
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
...
  /api/gps/source:
    put:
      operationId: GpsSourceConfigResource_updateGpsConfigSource
      tags:
      - "User: GPS Sources"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGpsSourceConfigDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
        "400":
          description: Bad Request
      summary: Update Gps Config Source
      security:
      - jwt:
        - USER
        - ADMIN
    get:
      operationId: GpsSourceConfigResource_getGpsSourceConfigs
      tags:
      - "User: GPS Sources"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Get Gps Source Configs
      security:
      - jwt:
        - USER
        - ADMIN
    post:
      operationId: GpsSourceConfigResource_addGpsSourceCnfig
      tags:
      - "User: GPS Sources"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGpsSourceConfigDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
        "400":
          description: Bad Request
      summary: Add Gps Source Cnfig
      security:
      - jwt:
        - USER
        - ADMIN
...
  /api/gps/source/{id}:
    delete:
      operationId: GpsSourceConfigResource_deleteGpsSourceConfig
      tags:
      - "User: GPS Sources"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Delete Gps Source Config
      security:
      - jwt:
        - USER
        - ADMIN
...
  /api/gpslogger:
    post:
      summary: Ingest GPS Logger location
      description: Receives a GPS Logger location update and stores it as a GPS point
        for the matching source token.
      operationId: GpsLoggerResource_handleGpsLogger
      tags:
      - "Public: GPS Integrations: GPS Logger"
      parameters:
      - name: X-Limit-D
        in: header
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
...
  /api/homeassistant:
    post:
      summary: Ingest Home Assistant location
      description: Receives a Home Assistant location update and stores it for the
        matching source token.
      operationId: HomeAssistantResource_handleHA
      tags:
      - "Public: GPS Integrations: Home Assistant"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HomeAssistantGpsData"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
...
info:
  title: GeoPulse API
  version: 1.34.1
  description: GeoPulse backend REST API documentation.
  contact:
    name: tess1o
    email: kerriden1@gmail.com
servers:
- url: http://localhost:8080